Dockerfile overview added.
Currently, the navigation areas left and right of the content are not visible to the reader when certain browser window widths are used (less than 1,000 pixels width).

The navigation tree doesn't come with any help about the different Dockerfile commands.

So, this page serves the purpose of a quick Dockerfile reference overview that can be pinned or bookmarked for easy reference.