Fix references to docker-compose command
Apologies if this is not the correct way to fix this. I just noticed on https://docs.docker.com/engine/reference/commandline/compose/ that the first example is using `docker-compose` instead of `docker compose`. It was very confusing trying to open a PR to fix it, as the "Edit this page" link leads into an (I think?) outdated repository, and the other option "Request docs changes" seems overly complicated for a simple copy&paste fix like this.