Mention the necessity of `--network host`
This example is, honestly, idiotic, because it doesn't even explain what task it solves.
No one needs to "just pass *_PROXY envs to created containers", as it is useless by itself.
People want to make containers have internet access via proxy, but the proxy is usually on the host machine's host or networks, not on docker containers networks!
The example described in the docs just tells how to configure containers query THEIR 127.0.0.1:3001, not host machine's 127.0.0.1:3001. This is useless!