Update environment-variables.md to include critical syntax info
I looked forever to find this information.  The expected section to find this information is this page, not where it ends up at the bottom of overview.

Based on the v3 overview of compose-files, it seems it can be included with this line- as thats how they did it there.

A link would also work, or a copy of the whole section.  Using include has the added benefit however of making a single section with changes updated across the docs if needed.