Fix Dockerfile COPY/ADD examples without a trailing slash
When copying files to a destination directory, the classic builder requires
the destination to have a trailing slash (to indicate the target is a _directory_.
not a filename). BuildKit is a bit more flexible in this, and will assume the
target is a directory, but users following the example with buildkit disabled
might see an error message, e.g.:

    Sending build context to Docker daemon  3.072kB
    Step 1/2 : FROM busybox
     ---> 69593048aa3a
    Step 2/2 : COPY *.go .
    When using COPY with more than one source file, the destination must be a directory and end with a /

It doesn't hurt to be explicit, so this patch updates some examples to prevent
this.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>