add explicit instruction to stop local server
This relates to the [Test the application section of the Build images tutorial](https://docs.docker.com/language/nodejs/build-images/#test-the-application). 

In this section, `node server.js` is used to start the server locally. The reader is not instructed to stop the server. 

In the [Run containers](https://docs.docker.com/language/nodejs/run-containers/) tutorial that follows this one, the reader is instructed to start up the container, and try to access `http://localhost:8000/test` which is now being served from the Node.js process inside the container. However, if the local Node.js process from the `node server.js` invocation earlier, they won't see the "Failed to connect to localhost port 8000" error which they're supposed to see. 

So this modification adds an explicit instruction to the "Build containers" tutorial to stop the `node server.js` process, to avoid any potential confusion in the "Run containers" tutorial.

Great tutorials, btw!