Mention about use of `|` or `&&` in one's commands
I think the root documentation should have a little bit more about what's allowed and what not in the `command`. I find it handy to place even complicated commands, within the YAML file (instead of spreading them into external `.sh` files), just like we sometimes do in `npm` `package.json`.

Anyhow, the fact that `|` or `&&` quietly chops my command was surprising.

The content of the PR is just a suggestion.