Dockerfile: try using official jekyll/jekyll:4 image
The starefossen image hasn't been maintained for some time, so perhaps
it's time to switch to the upstream jekyll image.

This image does not have some of the plugins installed that we use
(no github pages, and relative links), so we need to install these
through the Gemfile.

I'm experimenting here to see if we can preserve Jekyll's cache in
between builds (which is not exactly trivial), and some work is still
needed.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>