Add mentione of donotstart
This post on StackOverflow lead me to try `donotstart` profile. Confirmed, this setting on docker-compose 1.29.2 prevents container from spinning up. I do not see an error message, so I guess it is a feature.
Would be great to have this documented.

In case this is not documented for purpose, can you tell me why, so I can either reflect this in PR or close it?