Updated `## Persist the todo data` section
I provided a 3-step procedure from lines 81 to 102 on how to access the container's shell environment, since It was unclear how to access the `/etc/todos` directory. Also, nowhere, in the tutorial, was it mentioned that containers are *self-contained shell environments*. This information is crucial for beginners following along with the tutorial.

Take care.