Update garbage-collection.md
Based on the pull request distribution/distribution#2302 there is a parameter for cleaning up untagged manifests available. In the original pull request it is called "-m" / "--delete-manifests", but in the current version of registry and it's garbage-collector, it's called "--delete-untagged" (checked with "registry garbage-collect -h" inside the latest registry-image based container). Anyway, it was missing from the documentation so I thought it would be useful to add there.

Related issue and PR:
- https://github.com/distribution/distribution/issues/2301
- https://github.com/distribution/distribution/pull/2302