Move example of symbolic link closer to related text (#14228)
* Move example of symbolic link closer to related text

This is to co-locate the text explaining why a symbolic link might be required with the example of the symbolic link creation. This will make it more clear that the symbolic link example is related to the note immediately above.

* Insert line break prior to example

* Fix indentation of quote block

It now matches the indentation of the instructions code blocks and other quote blocks.