Update "docker run" command
The Dockerfile in get-started/02_our_app.md was updated to include the
installation of SQLite build dependencies (python2, g++ and make) to
make the guide compatible with non-amd64 computers.

The docker run command in get-started/06_bind_mounts.md was not updated,
so the command does not work as is. This commit brings the SQLite build
dependencies into the the docker run command as well.