[WIP] test engine cli reference changes
This updates the yamldocs for the engine cli reference with updated
sources that include anchor tags for flags; this should make the
flags in the "options" table clickable and scroll to the relevant
section on the page (where present).

This was built from https://github.com/docker/cli/pull/3509, which
is not yet merged (and has changes for the upcoming engine release,
so cannot yet be published here).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>