reformat "get started" 06 - bind-mounts
The "docker build" example was written as an in-line example, but directly
followed by a full-stop at the end of the sentence, which made it easy to
confuse the extra `.` to be part of the command to run.

This moves it to a code block to allow the example to be highlighted, and
less confusing.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>