Arm compatibility
The curl command was fetching an outdated version that no longer exists on the repository, at least for my system on a raspberry pi, so I changed that manually, although it probably will become outdated eventually as well.

I also added clarification for installing on arm, specifically that $(uname -m) fetches `arm71`-at least for me-when the curl command and existing repositories expect it to read `armv7`. I'm not sure how to automate that so until then I added a text box.