reference: remove "description" header, and use it for "extended description"
The "description" header included the short description of each command. In most
cases, this description was very short ("run a container"), and adding the extra
header before it only was adding extra noise.

This patch:

- removes the top "description" header
- renames the existing "extended description" header to "description" (a hidden
  "extended-description" anchor is added for backward compatibility)

As the extended description can be long (hopefully!), there may be a long distance
between the `Usage` section and the `Options` section. To help users navigate
to the list of available options, an extra line is printed if an extended
description is available for the command, including a link to the corresponding
section:

> Refer to the options section for an overview of available OPTIONS for this command.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>