engine: remove instructions for "nightly" and "test" channels
The "nightly" channel hasn't been updated for a long time, so while it's
not actively updated, we should not show instructions for it.

The "test" channel is still updated in preparation of new releases, but
including the instructions in between the regular installation procedure
introduced quite some "noise" that's irrelevant to most users (who would
be installing the "stable" channel).

This patch removes the instructions to enable the nightly and test channels
from the installation instructions.

While making these changes, I also synced somme wording between distros
for consistency, and removed a stray section about storage-drivers, as
that's already outlined in the "optional post-installation steps".

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>