engine: move note about package split to 18.09 release notes
containerd, docker engine and the docker cli were split into separate
packages in the 18.09 release. We kept this note on the release-notes
for current releases, but by now it's mostly "old news", so move the
warning to the 18.09 release notes page (which was the release in which
this change was made).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>