Add a note that local is a bash keyword
The example will run from the command line, but not as part of a bash script.  It took me an embarrassing several minutes to figure out why my start script wouldn't run with the cryptic error "docker: invalid reference format," so I thought a one-line note was worth making in the documentation.