extensions-sdk: remove "title" from metadata
These pages already contain a H1, so omit the title metadata to prevent
Jekyll from generating a second H1.

The downside of using the in-page H1 is that the reading-time is shown
before the H1. As these pages are for reference, and usually not read
"as a whole", this patch omits the reading-time indicator for these pages.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>