network/proxy: add a note pointing at docker engine and Docker Desktop
The title is "Configure Docker to use a proxy server" which is generic
and implies this is the page which describes everything.

However the scope of the instructions is limited to containers:

> If your container needs to use an HTTP, ...

There are other HTTP proxy cases, for example

- for Docker Desktop to log into Docker
- for pulling and pushing containers
- for docker scan

Add a note at the top of this page to clarify the scope and link to
instructions for other cases.

Signed-off-by: David Scott <dave.scott@docker.com>