desktop: settings: remove "-settings" suffix from URLs
The "-settings" suffix in the URL duplicated the "/settings/" directory
in the URL, which made it redundant.

Using "/mac/", "/windows/" and "/linux/" keeps the URLs shorter, more
memorable, and more consistent with other locations, making it easier
to "guess" the URL when navigating using the browser's location bar.

No redirects were added, because the "-settings" pages have not yet
been published.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>