Fix use of bullet list (#15090)
* Fix use of bullet list

Otherwise it would read like so:

To summarize, a container: Containers are isolated from each other and run their own software, binaries, and configurations.

Alternatively, to keep it as a bullet point:

To summarize, a container: 
...
is isolated from other containers and runs its own software, binaries, and configurations.

Both are correct. I can change the PR to the other option if preferred.

* Fix use of bullet list

Thanks, @craig-osterhout 
 
As requested, here's the commit for the alternative. Please reject the other commit and accept this one.