Add /go/linux-credentials/ redirect
Docker Desktop for Linux currently has a link to;
https://docs.docker.com/desktop/linux/#credentials-management

which has moved, and didn't have a redirect, causing the link to break;
unfortunately we can't fix existing versions of Docker Desktop; redirects
are being added (but won't be able to redirect to the correct anchor).

This patch adds a https://docs.docker.com/go/linux-credentials/ redirect,
so that we can continue using that URL even if content moves around.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>