DD4L: Replace download links with per-distro install instruction links
Currently it's easy for users to skip the per-distro installation steps,
since direct download links for our packages are presented first. This
can lead to unnecessary frustration, since DD will not install if
pre-requisite steps are not followed. So instead of presenting package
download links in the most prominent spot, point users to per-distro
installation instructions.

Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>