Update Linux instructions for Docker & Prometheus
The instructions for Linux do not work. Prometheus running in the container cannot access docker metrics on the host machine using `localhost:9323`, as localhost refers to the container itself.

Using Docker v 20.10 or greater, `host.docker.internal` will reference the host machine when `host.docker.internal:host-gateway` is passed in as a host mapping.

The only thing that needs to be checked is if using `0.0.0.0:9323` for `metrics_addr` will work for Windows and Mac platforms as well. I found that using `localhost:9323" and `127.0.0.1:9323` did not work on Linux.

/remove-lifecycle stale