Update build-images.md
I am just learning Docker and following the instructions. The Dockerfile that the tutorial would have you build does not work. It leads to error: Python3 not found. I tried editing in different versions of Python, changing the CMD line accordingly, and rebuilding, but it did not work. However if I substituted in the Dockerfile from the docs found here https://hub.docker.com/_/python/ then it does work. I'm suggesting this edit so that the error doesn't occur, and so that the documentation is consistent. I suspect some of these lines are redundant and didn't need changing; I'm not sure what caused the initial error or why this fixes it if I'm being perfectly honest.