compose: update env file based on actual behavior (#15759)
* compose: update env file based on actual behavior

There has been a LOT of work on this area in Compose recently.
As part of this, we discovered that the documentation has been
incorrect for some time. That is, the documented behavior has
often never been indicative of reality (across both v1 and v2).

We've standardized and cleaned things up, so the documentation
now reflects current behavior. The Compose Spec requires similar
corrections/reconciliation, but we're starting here to ensure
that the more user-facing docs are in a good place.

* Update compose/env-file.md

* Update compose/env-file.md

Co-authored-by: Allie Sadler <102604716+aevesdocker@users.noreply.github.com>