install: remove redundant & mismatched versions (#16476)
We list the supported versions and then have an info panel
noting our macOS support policy. At the end of this, it then
repeated the supported versions. However, this list was out
of sync with the one immediately preceding it. Rather than
fix it, I removed the sentence entirely as it's literally the bullet
point it's providing clarification on in the first place.

Signed-off-by: Milas Bowman <milas.bowman@docker.com>

Signed-off-by: Milas Bowman <milas.bowman@docker.com>