Note the bug in tmpfs permissions
There is an open bug in tmpfs that results in permissions being reset on container restart. At this point it has tripped up quite a few people. This adds a hint about the issue, and a link to a the workaround we ultimately used (a functionality not mentioned elsewhere in the docs).