Update run-tests.md
The tests literally didn't work for me. I'm not sure what breaking changes appear to have come in since they were first written, but a major problem was that calling `.WithNetwork()` passing in a not-yet-started network would throw an exception. Most of the changes are around coping with that - moving code from ctor to the `InitializeAsync()` method.