Update uninstall.md
I discovered a daemon file that was persisting from my old mac machine.  It had been restored on the new computer from a Time Machine backup and inherited many of the system configurations from the Intel macbook.  Here is how I fixed that uninstall.  You may want to investigate as the daemon file was not purged in any of the uninstall methods from the existing documentation.