compose: add upgrade notes for 2.17.x
Compose had several correctness changes in recent releases, which
have caught folks off-guard. Added an advisory "Upgrade Notes"
section to the v2.17.0 release section with the highlights.

I also removed a feature (`dockerfile_inline`) because there is
a bug in the initial implementation rendering it unusable for the
moment, so we will (re-)announce it once fixed.

Signed-off-by: Milas Bowman <milas.bowman@docker.com>