Fix typos and rewrite a few confusing phrases (#17807)
Fix typos and minor language corrections

Remove unnecessary hyphens

Rewrite a few confusing phrases -- "ease later changes" is not clear so I rephrased it

Remove the ellipsis on line 561 as the final parameter. It was probably added to indicate a variable number of parameters. However, I checked the link given to "Dockerfile reference for the CMD instruction" and there was no ellipsis there, so I see no reason for it to be here. I considered replacing it with three dots and spaces inbetween to make it more visible, but I think the ellipsis is not really necessary.