add "go" redirect for deprecated image formats
We currently print the full URL in warnings, and we may want to create a
better page for this at some point;

    docker pull docker:1.0.1
    1.0.1: Pulling from library/docker
    Image docker.io/library/docker:1.0.1 uses outdated schema1 manifest format. Please upgrade to a schema2 image for better future compatibility. More information at https://docs.docker.com/registry/spec/deprecated-schema-v1/

Adding a new "/go/" redirect, so that we can point it to the relevant
content as we go. I tried to keep the URl  somewhat short, so didn't
include "deprecated", but naming is hard, so input welcome.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>