ci: add Compose Watch configuration
This is a proof-of-concept that adds live update support while
developing on the docs by adding the experimental `watch` section.

The existing Compose configuration has been separated out into a
`compose.preview.yaml`, which builds the full site and serves it
from nginx to mirror production.

The new Compose configuration does a `jekyll serve` and has a couple
hacks for performance, namely that it does not run any of the local
plugins from `_plugins`. They're slow and run on every change but
aren't strictly necessary.