feat: add fs-based toc rendering
This change adds filesystem-based generation of the sidebar navigation.
It's an opt-in per top-level section, based on the existence of a
`legacytoc` front matter attribute. Setting the attribute to `false`
opts the page into using an fs-based rendering for the sidebar. This
rendering strategy should be used to opt-in an entire top-level section
at once, so the attribute should cascade from the top and down.

Signed-off-by: David Karlsson <35727626+dvdksn@users.noreply.github.com>