engine: remove logentries logging driver
The logentries logging driver was removed in Docker Engine v25. The logentries
service was already shut-down, so it's unlikely there's still active users of
this driver.

When upgrading to Docker Engine v25, any container configured with this driver
will automatically be migrated to the "local" logging driver, so no action
is required from the user, and we can safely remove these docs.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>