A '--dns' server will not be replaced by default DNS
Google's DNS servers are currently used by containers on the default
bridge network, when none of the host's /etc/resolv.conf nameservers
are usable (host loopback addresses that the container can't see).

But, if a server is supplied via the --dns option, it'll just appear
in the container's resolv.conf (on the default bridge) or as an
upstream server for the internal resolver.

So, fix the networking-intro page's description, noting that requests
to --dns servers happen in the container's namespace.

Signed-off-by: Rob Murray <rob.murray@docker.com>