build: incorporate build guide into manuals
The build guide that I once wrote now duplicates a lot of content with
the concept guide launched a few months back. It doesn't make sense to
cover the same topics multiple times, so I'm removing the older build
guide and incorporating some of the topics into the build manual
instead.

Signed-off-by: David Karlsson <35727626+dvdksn@users.noreply.github.com>