hugo: replace dots with dashes in content filenames
With the recent change in moving content to the /manuals section, and
manipulating content permalinks to use `/:sections[1:]/:slugorfilename`,
source filenames with dots in them did not receive a .html extension in
the site output directory. This seems to be a quirk (bug?) with the
`:slugorfilename` permalink token in Hugo.

To work around this, I have removed the dots in markdown filenames and
replaced them with dashes.

Signed-off-by: David Karlsson <35727626+dvdksn@users.noreply.github.com>