search: don't hide search drop-down on focus change
Focus/blur events are handled diffrently in Safari compared to other
browsers. For a better experience, let's not attempt to close the search
results drop-down if focus exits the search bar area. We still have a
click.outside and a keyup listener for "escape" on window to handle the
close, and a "close" button in the input bar (which is now also
focusable).

Signed-off-by: David Karlsson <35727626+dvdksn@users.noreply.github.com>