site: automatically scroll down the sidebar to reveal the current page
Some pages are nested quite deep in the sidebar, which means they aren't
automatically visible on pageload unless the sidebar is scrolled. This
commit adds a scrolldown function to the sidebar to check if the element
representing the current page is in view. If it isn't, the sidebar is
scrolled down to a position where the current page comes into view.

Signed-off-by: David Karlsson <35727626+dvdksn@users.noreply.github.com>