Update _index.md about switching container mode (#21614)
<!--Delete sections as needed -->

## Description

<!-- Tell us what you did and why -->
Docker desktop was installed by an automated system on my device and
when I went to change the WSL Integration settings, the panel was not
available. The app had been configured to use "Windows containers mode"
and I had to switch it to "Linux containers mode" to be able to see the
panel. I have added text to help others who may face the same problem.

## Related issues or tickets

<!-- Related issues, pull requests, or Jira tickets -->

## Reviews

<!-- Notes for reviewers here -->
<!-- List applicable reviews (optionally @tag reviewers) -->

- [ ] Technical review
- [ ] Editorial review
- [ ] Product review

---------

Co-authored-by: Allie Sadler <102604716+aevesdocker@users.noreply.github.com>