release rollouts (#22044)
<!--Delete sections as needed -->

## Description

Went through Hotjar feedback and we get quite a few reports of the
latest version not being immediately available. (We also get this
question in GitHub and Kapa and on Slack). Have added a line to explain
this.

Also removed the link to the public roadmap as I'm not sure how often
this is updated

## Related issues or tickets

<!-- Related issues, pull requests, or Jira tickets -->

## Reviews

<!-- Notes for reviewers here -->
<!-- List applicable reviews (optionally @tag reviewers) -->

- [ ] Technical review
- [ ] Editorial review
- [ ] Product review