feat: add testcontainers-oss quickstart guides (#22120)
- **chore: add testcontainers entry in the OSS section of Manuals**
- **chore: adds quickstarts for tc-go and tc-java**

<!--Delete sections as needed -->

## Description
This PR adds the Testcontainers OSS section in the Manuals > OSS section
of the docs.

It's adding the following pages:

- index page with a basic description of the project, with two cards:
  - why use testcontainers, linking to a benefits page
  - quicktart, linking to a folder with two pages: Go and Java

Below the cards grid, it links to the testcontainers.com/getting-started
page.

The benefits page includes some of the content of the above
getting-started page, and the Go and Java quickstart pages are copying
the quickstart pages from Go
(https://golang.testcontainers.org/quickstart/) and Java
(https://java.testcontainers.org/quickstart/) in its 3 flavours (junit4,
junit5 and spock)

<!-- Tell us what you did and why -->

## Related issues or tickets

<!-- Related issues, pull requests, or Jira tickets -->

## Reviews

<!-- Notes for reviewers here -->
<!-- List applicable reviews (optionally @tag reviewers) -->

- [x] Technical review
- [ ] Editorial review
- [x] Product review

---------

Co-authored-by: aevesdocker <allie.sadler@docker.com>