Rails guide upgrade (part 2) (#22114)
## Description

This PR has two goals. First of all it continues work that has been done
in #21559. Besides this, and this is more significant, I tried to
improve the structure of the documentation that is currently used for
different language-specific guides like
[PHP](https://docs.docker.com/guides/php/configure-ci-cd/),
[Go](https://docs.docker.com/guides/golang/),
[Python](https://docs.docker.com/guides/python/) and others, including,
of course, Ruby itself.

Each of these guides currently has a [Configure
CI/CD](https://docs.docker.com/guides/python/configure-ci-cd/) section.
Inside this section there is a GitHub Actions workflow example that has
nothing in common with a CI/CD pipeline. It's just an example of how to
build and push an image to a Docker Hub registry. We should be clear in
our documentation and not mislead our users. This was the main reason
why I renamed this section to "Automate your builds with GitHub
Actions". I also updated the content of this section to reflect the new
name and to make it more clear what the user can expect from this guide.
I suggest the same be done for all other language-specific guides.

Besides this, I changed the order of the sections in the Ruby guide. The
"Develop your app" section has been moved down to the bottom of the
guide. This makes more sense to me because of two reasons:

1. It is more important to start using Docker Hub right after you added
the Dockerfile to your project (section number one in all
language-specific guides).

2. I can hardly imagine anybody using Docker to run the app locally for
development purposes (at least for Ruby). What is really essential and
useful is to know how to run services, that are required by your app,
like a database, a cache server, or a local LLM. This is why the
"Develop your app" section should be rewritten to explain how to run the
infrastructure services that are required by the app and not the app
itself.

Below there are screenshots reflecting the changes that have been made
in this PR.

**Before**
<img width="1512" alt="Screenshot 2025-02-27 at 11 26 39"
src="https://github.com/user-attachments/assets/1ca06aea-ffeb-4efb-a14d-27254d2a2110"
/>

**After** 
<img width="1512" alt="Screenshot 2025-02-27 at 11 26 01"
src="https://github.com/user-attachments/assets/7abbe8b7-d1b3-480f-8105-49f967b51e47"
/>

## Related issues or tickets

#21559

## Reviews

- [x] Technical review
- [x] Editorial review
- [ ] Product review