docs: added note on populating builds tab when building windows container images (#22389)
<!--Delete sections as needed -->

## Description

<!-- Tell us what you did and why -->
If a user switches to Windows containers and builds an image using the
`docker build` command as it (using the legacy builder), the Builds tab
is not populated. This is because the build history is provided by
Buildkit. To populate the Builds tab, the customer must either set
`DOCKER_BUILDKIT=1` or use the `buildx build` command.

Current documentation does not mention this and a support ticket was
raised.

## Related issues or tickets

<!-- Related issues, pull requests, or Jira tickets -->
- SEG-1116 (Slack thread linked in ticket for context)

## Reviews

<!-- Notes for reviewers here -->
<!-- List applicable reviews (optionally @tag reviewers) -->

- [X] Technical review @karman-docker 
- [X] Editorial review
- [ ] Product review

---------

Co-authored-by: Allie Sadler <102604716+aevesdocker@users.noreply.github.com>