Update Laravel guide (bump to Laravel 12 and php 8.4) (#22699)
<!--Delete sections as needed -->

## Description

Introduces small update of the Laravel framework guide (bump versions to
the latest Laravel 12 and php 8.4 + updates the related examples repo -
the fork needs to be synced).

## Related issues or tickets

#21268 

## Reviews

<!-- Notes for reviewers here -->
As was discussed on the last docs meeting, to simplify the process the
review will be sent to the one of Docker Captains first.
@igor-alexandrov @ajeetraina 

- [ ] Technical review
- [ ] Editorial review
- [ ] Product review