services: add notice of broken feature
Waiting for healthy dependencies is not supported in version 3 of docker compose, as explained in:
https://docs.docker.com/compose/releases/release-notes/#1100

The `version` key is also ignored now, so this whole feature is actually broken. I think at the very least, this should be mentioned in the documentation.

This took me 2 days of debugging to figure out, and a notice here would've saved me a lot of time.