Fix example usage of `COMPOSE_ENV_FILES` (#23253)
## Description
The example results in the second file being a command to be executed
and not part of the environment variable's value.

Quoting the value would address it as well, but thought it far more
likely the space was an accident rather than trying to show an example
of files containing spaces.