ENGDOCS-2981 (#23377)
<!--Delete sections as needed -->

## Description

Adds new embedded PAC admin setting (and missing `pac` setting)

Note: Needs to also be added to the settings reference page, but this
will be done after the page has been updated as there a few settings
missing and info on what's available in the console vs. the json script.
This will be a fast-follow after the release on Thursday

## Related issues or tickets

<!-- Related issues, pull requests, or Jira tickets -->

## Reviews

<!-- Notes for reviewers here -->
<!-- List applicable reviews (optionally @tag reviewers) -->

- [ ] Technical review
- [ ] Editorial review
- [ ] Product review